# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Public read and admin write contract for Throne & Liberty Ranked Wargames.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: getPlayers
      tags: [public]
      summary: Get public player directory
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Public player directory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerSummary"
  /players/{playerId}:
    get:
      operationId: getPlayer
      tags: [public]
      summary: Get a player profile
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
        "404":
          description: Player not found
  /matches:
    get:
      operationId: getMatches
      tags: [public]
      summary: Get recent completed matches
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        "200":
          description: Recent matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchSummary"
  /visitor-events:
    post:
      operationId: recordVisitor
      tags: [public]
      summary: Record an anonymous page visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitorEventInput"
      responses:
        "204":
          description: Visit recorded
        "429":
          description: Too many visit events
  /applications:
    post:
      operationId: createApplication
      tags: [applications]
      summary: Submit a six-player roster application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationReceipt"
        "400":
          description: Invalid roster application
        "429":
          description: Too many application attempts
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get admin analytics summary
      responses:
        "200":
          description: Admin summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
        "401":
          description: Admin authentication required
...
  /admin/classes:
    get:
      operationId: getAdminClasses
      tags: [admin]
      summary: List all class catalog entries
      responses:
        "200":
          description: Complete class catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassCatalogEntry"
    post:
      operationId: createClass
      tags: [admin]
      summary: Add a class to the catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassCatalogCreate"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassCatalogEntry"
  /admin/classes/{classKey}:
    patch:
      operationId: updateClass
      tags: [admin]
      summary: Edit, enable, or disable a class
      parameters:
        - name: classKey
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassCatalogUpdate"
      responses:
        "200":
          description: Class updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassCatalogEntry"
    delete:
      operationId: deleteClass
      tags: [admin]
      summary: Permanently delete an unused disabled class
      parameters:
        - name: classKey
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditedAction"
      responses:
        "204": { description: Class deleted }
        "409": { description: Class is referenced by historical records }
  /admin/applications:
    get:
      operationId: getAdminApplications
      tags: [admin]
      summary: List complete roster applications and their audit history
      responses:
        "200":
          description: Roster applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminApplication"
        "401":
          description: Admin authentication required
  /admin/applications/{applicationId}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [admin]
      summary: Approve, reject, or cancel a roster application
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminApplication"
        "401":
          description: Admin authentication required
        "404":
          description: Application not found
  /admin/matches:
    get:
      operationId: getAdminMatches
      tags: [admin]
      summary: List completed and discarded archive matches with audit history
      responses:
        "200":
          description: Archive matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminMatch"
        "401":
          description: Admin authentication required
    post:
      operationId: commitMatch
      tags: [admin]
      summary: Commit a verified scoreboard result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Match committed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchSummary"
        "401":
          description: Admin authentication required
        "400":
          description: Invalid match payload
  /admin/matches/{matchId}/scoreboard:
    get:
      operationId: getMatchScoreboard
      tags: [admin]
      summary: Open the retained scoreboard for a committed match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to a temporary scoreboard image URL
        "401":
          description: Admin authentication required
        "404":
          description: Match scoreboard not found
  /admin/matches/{matchId}:
    patch:
      operationId: correctMatch
      tags: [admin]
      summary: Correct a match while retaining its prior state in the audit history
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchCorrection"
      responses:
        "200":
          description: Corrected match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMatch"
        "401":
          description: Admin authentication required
        "404":
          description: Match not found
    delete:
      operationId: discardMatch
      tags: [admin]
      summary: Discard a match from public results while retaining its history
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditedAction"
      responses:
        "200":
          description: Discarded match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMatch"
        "401":
          description: Admin authentication required
        "404":
          description: Match not found
  /admin/matches/{matchId}/restore:
    post:
      operationId: restoreMatch
      tags: [admin]
      summary: Restore a discarded match to public results and rankings
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditedAction"
      responses:
        "200":
          description: Restored match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMatch"
        "401":
          description: Admin authentication required
        "404":
          description: Match not found
components:
  schemas:
...
    ClassCatalogEntry:
      type: object
      properties:
        key: { $ref: "#/components/schemas/ClassKey" }
        displayName: { type: string }
        aliases:
          type: array
          items: { type: string }
        active: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [key, displayName, aliases, active, sortOrder, createdAt, updatedAt]
    ClassCatalogCreate:
      type: object
      properties:
        key: { $ref: "#/components/schemas/ClassKey" }
        displayName: { type: string, minLength: 1, maxLength: 80 }
        aliases: { type: array, items: { type: string }, maxItems: 20 }
        sortOrder: { type: integer, minimum: 0, maximum: 10000 }
        actor: { type: string, minLength: 2, maxLength: 80 }
        reason: { type: string, minLength: 2, maxLength: 500 }
      required: [key, displayName, aliases, actor, reason]
    ClassCatalogUpdate:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 80 }
        aliases: { type: array, items: { type: string }, maxItems: 20 }
        active: { type: boolean }
        sortOrder: { type: integer, minimum: 0, maximum: 10000 }
        actor: { type: string, minLength: 2, maxLength: 80 }
        reason: { type: string, minLength: 2, maxLength: 500 }
      required: [actor, reason]
    TeamColor:
      type: string
      enum: [BLUE, RED]
...
    MomentumPoint:
      type: object
      properties:
        label:
          type: string
        blue:
          type: integer
        red:
          type: integer
      required: [label, blue, red]
...
    PlayerSummary:
      type: object
      properties:
        id:
          type: string
        characterName:
          type: string
        archetype:
          type: string
        matches:
          type: integer
        wins:
          type: integer
        winRate:
          type: number
        avgKills:
          type: number
        avgDamage:
          type: number
        form:
          type: string
      required: [id, characterName, archetype, matches, wins, winRate, avgKills, avgDamage, form]
    PlayerProfile:
      allOf:
        - $ref: "#/components/schemas/PlayerSummary"
        - type: object
          properties:
            avgAssists:
              type: number
            avgHealing:
              type: number
            records:
              type: array
              items:
                $ref: "#/components/schemas/PersonalRecord"
            weaponBreakdown:
              type: array
              items:
                $ref: "#/components/schemas/ClassSummary"
          required: [avgAssists, avgHealing, records, weaponBreakdown]
    PersonalRecord:
      type: object
      properties:
        label:
          type: string
        value:
          type: number
        achievedAt:
          type: string
      required: [label, value, achievedAt]
...
    MatchSummary:
      type: object
      properties:
        id:
          type: string
        matchDate:
          type: string
        winningTeam:
          $ref: "#/components/schemas/TeamColor"
        blueScore:
          type: integer
        redScore:
          type: integer
        note:
          type: string
        hasScreenshot:
          type: boolean
        participants:
          type: array
          items:
            $ref: "#/components/schemas/MatchParticipant"
      required: [id, matchDate, winningTeam, blueScore, redScore, note, hasScreenshot, participants]
...
    VisitorEventInput:
      type: object
      properties:
        path:
          type: string
          minLength: 1
          maxLength: 160
        referrer:
          type: string
          maxLength: 200
      required: [path]
    ApplicationInput:
      type: object
      properties:
        groupName:
          type: string
          minLength: 2
          maxLength: 80
        captainName:
          type: string
          minLength: 2
          maxLength: 40
        captainContact:
          type: string
          minLength: 3
          maxLength: 120
        notes:
          type: string
          maxLength: 1000
        submissionToken:
          type: string
          minLength: 16
          maxLength: 100
        formStartedAt:
          type: integer
          minimum: 0
        website:
          type: string
          maxLength: 200
        members:
          type: array
          minItems: 6
          maxItems: 6
          items:
            $ref: "#/components/schemas/ApplicationMemberInput"
      required: [groupName, captainName, captainContact, submissionToken, formStartedAt, website, members]
    ApplicationReceipt:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        message:
          type: string
      required: [id, status, message]
...
    ApplicationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, CANCELLED]
...
    AdminApplication:
      type: object
      properties:
        id:
          type: string
        groupName:
          type: string
        captainName:
          type: string
        captainContact:
          type: string
        notes:
          type: string
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        createdAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/AdminApplicationMember"
        audit:
          type: array
          items:
            $ref: "#/components/schemas/AuditRecord"
      required: [id, groupName, captainName, captainContact, notes, status, createdAt, members, audit]
    ApplicationStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [APPROVED, REJECTED, CANCELLED]
        actor:
          type: string
          minLength: 2
          maxLength: 80
        reason:
          type: string
          minLength: 2
          maxLength: 500
      required: [status, actor, reason]
    AuditedAction:
      type: object
      properties:
        actor:
          type: string
          minLength: 2
          maxLength: 80
        reason:
          type: string
          minLength: 2
          maxLength: 500
      required: [actor, reason]
...
    MatchInput:
      type: object
      properties:
        actor:
          type: string
          minLength: 2
          maxLength: 80
        matchDate:
          type: string
        winningTeam:
          $ref: "#/components/schemas/TeamColor"
        note:
          type: string
        screenshot:
          type: object
          properties:
            name:
              type: string
              minLength: 1
              maxLength: 255
            contentType:
              type: string
              enum: [image/png, image/jpeg, image/webp]
            base64:
              type: string
              minLength: 1
          required: [name, contentType, base64]
        participants:
          type: array
          minItems: 12
          maxItems: 12
          items:
            $ref: "#/components/schemas/MatchParticipantInput"
      required: [actor, winningTeam, screenshot, participants]
...
    AdminMatch:
      allOf:
        - $ref: "#/components/schemas/MatchSummary"
        - type: object
          properties:
            status:
              type: string
              enum: [COMPLETED, DISCARDED]
            audit:
              type: array
              items:
                $ref: "#/components/schemas/AuditRecord"
          required: [status, audit]
    AdminSummary:
      type: object
      properties:
        visitorsToday:
          type: integer
        visitorsThisWeek:
          type: integer
        applicationQueue:
          type: integer
        ocrConfidence:
          type: [number, "null"]
        activity:
          type: array
          items:
            $ref: "#/components/schemas/MomentumPoint"
      required: [visitorsToday, visitorsThisWeek, applicationQueue, ocrConfidence, activity]